+++ app/apps/dev_router.rb
@@ -8,14 +8,18 @@ class DevRouter
     @category = "Sandbox Apps"
   end
 
-  def self.apps(owner: OodSupport::Process.user.name, require_manifest: true)
+  # Get array of apps for specified owner
+  #
+  # @param owner [String] username of user to get apps for
+  # @return [Array<OodApp>] all apps owner has access to in sandbox
+  def self.apps(owner: OodSupport::Process.user.name)
     target = base_path(owner: owner)
     if target.directory? && target.executable? && target.readable?
-      target.children.map { |d|
-        ::OodApp.new(self.new(d.basename, owner))
-      }.select { |d|
-        d.valid_dir? && d.accessible? && (!require_manifest || d.manifest.valid?)
-      }
+      target.children.map { |d| OodApp.new self.new(d.basename, owner) }
+        .select(&:directory?)
+        .select(&:accessible?)
+        .reject(&:hidden?)
+        .reject(&:backup?)
     else
       []
     end